final changes

+++ WebService/WebService/Services/ProductService.cs
@@ -23,6 +23,7 @@ namespace WebService.Services
         private readonly IMongoCollection<Product> _productCollection;
         private readonly IMongoCollection<Product_List> _productListCollection;
         private readonly IMongoCollection<User> _userCollection;
+        private readonly IMongoCollection<Order> _orderCollection;
         private readonly INotificationService _notificationService;
 
         // define mongodb collection name
@@ -35,6 +36,7 @@ namespace WebService.Services
             _productCollection = mongoDatabase.GetCollection<Product>(CollectionName);
             _productListCollection = mongoDatabase.GetCollection<Product_List>("product_list");
             _userCollection = mongoDatabase.GetCollection<User>("user");
+            _orderCollection = mongoDatabase.GetCollection<Order>("order");
 
             _notificationService = notificationService;
         }
@@ -57,6 +59,39 @@ namespace WebService.Services
         {
             // Fetch all products
             var products = await _productCollection.Find(_ => true).ToListAsync();
+
+            // Fetch all product lists
+            var productLists = await _productListCollection.Find(_ => true).ToListAsync();
+
+            // Fetch all product lists
+            var venodrs = await _userCollection.Find(_ => true).ToListAsync();
+
+            // Join products with product lists to get product list names
+            var productListDictionary = productLists.ToDictionary(pl => pl.Id);
+
+            var vendorDictionary = venodrs.ToDictionary(pl => pl.Id);
+
+            foreach (var product in products)
+            {
+                if (productListDictionary.TryGetValue(product.Product_idProductList ?? string.Empty, out var productList))
+                {
+                    product.ProductListName = productList.Name;
+                }
+
+                if (vendorDictionary.TryGetValue(product.Product_idVendor ?? string.Empty, out var user))
+                {
+                    product.ProductVendorName = user.First_Name + " " + user.Last_Name;
+                }
+            }
+
+            return products;
+        }
+
+        //get all product 
+        public async Task<List<Product>> GetActiveProduct()
+        {
+            // Fetch all products
+            var products = await _productCollection.Find(p => p.IsActive == true).ToListAsync();
 
             // Fetch all product lists
             var productLists = await _productListCollection.Find(_ => true).ToListAsync();
@@ -118,8 +153,19 @@ namespace WebService.Services
         //delete product
         public async Task RemoveProduct(string id)
         {
+            // Check if any orders contain the product with status == 0 (pending)
+            var pendingOrders = await _orderCollection.Find(order => order.OrderItems.Any(item => item.ProductId == id) && order.Status == 0).ToListAsync();
+
+            if (pendingOrders.Any())
+            {
+                // Handle the case where there are pending orders
+                throw new Exception("Product cannot be deleted as there are pending orders.");
+            }
+
+            // Proceed with product deletion if no pending orders are found
             await _productCollection.DeleteOneAsync(x => x.Id == id);
         }
+
 
         //change product status
         public async Task ChangeProductStatus(string id)
